updated utils - mutations and queries
