--- client/src/utils/GlobalState.jsx
@@ -1,21 +1,0 @@
-import { createContext, useContext, useReducer } from "react";
-import { reducer } from './reducers'
-
-const StoreContext = createContext();
-const { Provider } = StoreContext;
-
-const StoreProvider = ({ value = [], ...props }) => {
-  const [state, dispatch] = useReducer(reducer, {
-    products: [],
-    categories: [],
-    currentCategory: '',
-  });
-
-  return <Provider value={[state, dispatch]} {...props} />;
-};
-
-const useStoreContext = () => {
-  return useContext(StoreContext);
-};
-
-export { StoreProvider, useStoreContext };
