--- client/src/utils/actions.js
@@ -1,4 +1,0 @@
-export const UPDATE_PRODUCTS = "UPDATE_PRODUCTS";
-
-export const UPDATE_CATEGORIES = "UPDATE_CATEGORIES";
-export const UPDATE_CURRENT_CATEGORY = "UPDATE_CURRENT_CATEGORY";
