+++ client/src/utils/mutations.js
@@ -1,33 +1,44 @@
 import { gql } from '@apollo/client';
 
-export const GET_USER = gql`
-  query GetUser($userId: ID!) {
-    getUser(userId: $userId) {
+export const CREATE_USER = gql`
+  mutation CreateUser(
+    $firstName: String!
+    $lastName: String!
+    $email: String!
+    $password: String!
+    $birthday: Date
+  ) {
+    createUser(
+      firstName: $firstName
+      lastName: $lastName
+      email: $email
+      password: $password
+      birthday: $birthday
+    ) {
+      id
+      firstName
+      lastName
+      email
+      birthday
+    }
+  }
+`;
+
+
+
+export const CREATE_BOOKING = gql`
+  mutation CreateBooking($userId: ID!, $services: [ServiceInput]!, $staffId: ID!, $date: String!, $time: String!) {
+    createBooking(userId: $userId, services: $services, staffId: $staffId, date: $date, time: $time) {
       _id
     }
   }
 `;
 
-export const GET_STAFF = gql`
-  query GetStaff {
-    getStaff {
-      _id
+export const LOGIN = gql`
+  mutation Login($email: String!, $password: String!) {
+    login(email: $email, password: $password) {
+      token
     }
   }
 `;
 
-export const GET_SERVICES = gql`
-  query GetServices {
-    getServices {
-      _id
-    }
-  }
-`;
-
-export const GET_BOOKINGS = gql`
-  query GetBookings {
-    getBookings {
-      _id
-    }
-  }
-`;
