+++ client/src/utils/queries.js
@@ -1,25 +1,25 @@
 import { gql } from '@apollo/client';
 
-export const CREATE_USER = gql`
-  mutation CreateUser($name: String!, $email: String!, $password: String!) {
-    createUser(name: $name, email: $email, password: $password) {
+export const GET_STAFF = gql`
+  query GetStaff {
+    getStaff {
       _id
     }
   }
 `;
 
-export const CREATE_BOOKING = gql`
-  mutation CreateBooking($userId: ID!, $services: [ServiceInput]!, $staffId: ID!, $date: String!, $time: String!) {
-    createBooking(userId: $userId, services: $services, staffId: $staffId, date: $date, time: $time) {
+export const GET_SERVICES = gql`
+  query GetServices {
+    getServices {
       _id
     }
   }
 `;
 
-export const LOGIN = gql`
-  mutation Login($email: String!, $password: String!) {
-    login(email: $email, password: $password) {
-      token
+export const GET_BOOKINGS = gql`
+  query GetBookings {
+    getBookings {
+      _id
     }
   }
 `;
