+++ server/schemas/typeDefs.js
@@ -1,7 +1,3 @@
-// schemas/typeDefs.js
-
-const { gql } = require('apollo-server-express');
-
 const typeDefs = gql`
 # schemas/typeDefs.js
